# docs/swagger.yaml
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Soand API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Soand API
  version: "1.0"
paths:
...
  /comments/:comment_id:
    delete:
      description: Deletes a specific comment for the authenticated user
      parameters:
      - description: Comment ID (MongoDB ObjectID)
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid comment ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not delete comment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
    patch:
      consumes:
      - application/json
      description: Updates the text of a specific comment for the authenticated user
      parameters:
      - description: Comment ID (MongoDB ObjectID)
        in: path
        name: comment_id
        required: true
        type: string
      - description: New comment text
        in: body
        name: comment
        required: true
        schema:
          properties:
            new_text:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid comment ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not update comment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - comments
  /comments/:post_id:
    get:
      description: Retrieves a paginated list of comments for a specific post
      parameters:
      - description: Post ID (MongoDB ObjectID)
        in: path
        name: post_id
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: string
      - description: 'Number of comments per page (default: 10)'
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of comments
          schema:
            items: {}
            type: array
        "400":
          description: Invalid post ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not fetch comments
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get comments by post ID
      tags:
      - comments
  /comments/react:
    post:
      consumes:
      - application/json
      description: Adds or updates a reaction to a specific comment
      parameters:
      - description: Comment ID to react to
        in: query
        name: comment_id
        required: true
        type: string
      - description: Reaction details
        in: body
        name: reaction
        required: true
        schema:
          properties:
            type:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Reaction successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid comment ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not process reaction
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: React to a comment
      tags:
      - comments
  /comments/ws:
    get:
      consumes:
      - application/json
      description: Establishes a WebSocket connection for real-time comment updates
        on a specific post
      parameters:
      - description: Post ID to subscribe to comments for
        in: query
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Missing or invalid post ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: WebSocket upgrade failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: WebSocket connection for real-time comments
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    description: '"JWT Authorization header using the Bearer scheme. Example: ''Bearer
      {token}''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"